+++ caliko-visualisation/src/main/java/au/edu/federation/caliko/visualisation/Entity.java
